+++ Changelog.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-8972: fixed display of select componet with custom dataSrc on Data Tab
  - FIO-8954: fixed previously selected option is displayed in metadata with data source = url
  - FIO-8986: fixed setting up of default value with hidden fields
+ - FIO-8938: fixed label display for multiple select with dataSrc resource
 
 ### New Features